[refactor] : 파일 이름 변경, props 수정
