+++ app/auth/login/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useRouter } from 'next/navigation';
-import Button from '@/components/ui/button';
+import Button from '@/components/ui/BaseButton';
 import Logo from '@/components/icons/Logo';
 import { FaGithub } from 'react-icons/fa';
 import { FcGoogle } from 'react-icons/fc';
