+++ app/page.tsx
@@ -1,9 +1,7 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
 import LandingPage from './LandingPage';
 
-export default function Home() {
-  const router = useRouter();
-  return <LandingPage onNavigate={() => router.push('/auth/login')} />;
+export default function Page() {
+  return <LandingPage />;
 }
